Updated unit tests
