+++ test/views/html/pages/CIAPIViews/NationalMinimumWageCUIViewSpec.scala
@@ -32,7 +32,7 @@ class NationalMinimumWageCUIViewSpec extends ChatViewBehaviours with Matchers wi
     "rendered" must {
       behave like normalCuiPage(
         createView,
-        "Ask HMRC online - Ask HMRC - GOV.UK",
+        "Chat with HMRC’s digital assistant about tax - GOV.UK",
         "Ask HMRC online"
       )
     }
