+++ test/views/html/pages/testOnly/persistChatPageOneViewPopupSpec.scala
@@ -32,7 +32,7 @@ class persistChatPageOneViewPopupSpec extends ChatViewBehaviours with Matchers w
     "rendered" must {
       behave like normalCuiPage(
         createView,
-        "Persist Chat Page One Popup - Ask HMRC - GOV.UK",
+        "Chat with HMRC’s digital assistant about tax - GOV.UK",
         "Persist Chat Page One Popup"
       )
     }
